+++ tenpy/tools/misc.py
@@ -348,7 +348,7 @@ def any_nonzero(params, keys, verbose_msg=None):
             val = params.get(k, None)
             if val is not None and np.any(np.array(val) != 0.):  # count `None` as zero
                 if verbose:
-                    print(verbose_mesg)
+                    print(verbose_msg)
                     print(str(k) + " has nonzero entries")
                 return True
     return False